cloud_run/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
cloudevents==1.10.1
google-cloud-bigquery
google-cloud-storage